+++ backend/controllers/mencari.kerja.controller.js
@@ -1,8 +1,9 @@
 import MencariKerja from "../models/mencari.kerja.model.js";
 import User from "../models/user.models.js";
-
+import fs from "fs";
+ 
 export const createMencariKerja = async (req, res) => {
-  try {
+  try { 
     const userId = req.params.userId;
     const { pendidikan_akhir, alamat, alasan, email } = req.body;
     const user = await User.findById(userId);
@@ -11,7 +12,7 @@ export const createMencariKerja = async (req, res) => {
         status: false,
         message: "Data user tidak ditemukan",
       });
-    } 
+    }
 
     const mencariKerja = new MencariKerja({
       user: userId,
@@ -22,6 +23,10 @@ export const createMencariKerja = async (req, res) => {
       alamat: alamat,
       alasan: alasan,
       jenis: "Mencari Kerja",
+      gambar: req.file ? req.file.path : null,
+      urlGambar: req.file
+        ? `${req.protocol}://${req.get("host")}/${req.file.path}`
+        : null,
     });
 
     const cekData = await MencariKerja.findOne({ user: user });
@@ -32,12 +37,11 @@ export const createMencariKerja = async (req, res) => {
       });
     }
 
-    mencariKerja.save();
-
+    const result = mencariKerja.save();
     return res.status(201).json({
       status: true,
       message: "Data berhasil ditambahkan",
-      data: mencariKerja,
+      data: result,
     });
   } catch (error) {
     return res.status(403).json({
@@ -98,11 +102,33 @@ export const updateMencariKerja = async (req, res) => {
     const userId = req.params.userId;
     const { nama, angkatan, pendidikan_akhir, alamat, alasan, email } =
       req.body;
-    const mencariKerja = await MencariKerja.findOneAndUpdate(
+
+    let mencariKerja = await MencariKerja.findOneAndUpdate(
       { user: userId },
       { nama, angkatan, pendidikan_akhir, alamat, alasan, email },
       { new: true }
     );
+
+    if (!mencariKerja) {
+      return res.status(401).json({
+        status: false,
+        message: "Data tidak ditemukan",
+      });
+    }
+
+    if (req.file && mencariKerja.gambar) {
+      fs.unlinkSync(mencariKerja.gambar);
+    }
+
+    // Jika ada file gambar yang diupload, perbarui juga field gambar dan urlGambar
+    if (req.file) {
+      mencariKerja.gambar = req.file.path;
+      mencariKerja.urlGambar = `${req.protocol}://${req.get("host")}/${req.file.path}`;
+    }
+
+    mencariKerja = await mencariKerja.save();
+
+
     return res.status(201).json({
       status: true,
       message: "Data berhasil diupdate",
@@ -119,11 +145,31 @@ export const updateMencariKerja = async (req, res) => {
 export const deleteMencariKerja = async (req, res) => {
   try {
     const userId = req.params.userId;
-    const mencariKerja = await MencariKerja.deleteOne({ user: userId });
+
+    const mencariKerja = await MencariKerja.findOne({ user: userId });
+
+    if (!mencariKerja) {
+      return res.status(401).json({
+        status: false,
+        message: "Data tidak ditemukan",
+      });
+    }
+
+     // Langkah 2: Hapus gambar fisik jika ada
+     if (mencariKerja.gambar) {
+      fs.unlinkSync(mencariKerja.gambar);
+    }
+
+    const result = await MencariKerja.deleteOne({ user: userId });
+
+    if (result.deletedCount === 0) {
+      return res.status(404).json({ error: "Data kerja tidak ditemukan." });
+    }
+
     return res.status(202).json({
       status: true,
       message: "Data berhasil di hapus",
-      data: mencariKerja,
+      data: result,
     });
   } catch (error) {
     return res.status(402).json({
